Enter commit message.  Lines beginning with 'HG:' are removed.  Leave message empty to abort commit.  --  user: sk.anirudh  branch 'default'  added simulateWirelessNetwork.cc
I added simulateWirelessNetwork.cc which simulates a VANET, with the appopriate power of 33 dbm and the minimum data rate. There is still an issue with the real time deadline with simulations aborting pretty quickly every now and then.
Most of the problem is due to the fact that all nodes have to be initialized at time t=0, and as the number of nodes grow, this becomes exceedingly harder. I am going to try and remedy this problem by scheduling and staggering node initialisation.