This was a first attempt to try and set up a server in the cloud. The code does not work, but I need to document progress, so here goes. I next tried porting over the whole thing to the emulation framework which is tap-wimax.cc which does not work either.
The issues are two fold. In wimax-server.cc it seems impossible to configure static routes in simulation since ns3 allows this only for p2p and csma lans.
Besides, even if we did manage to do it, it would be useless since we need to move it to an emulated framework in any case.

In the emulated framework, I can't even get a basic ping to work and the code keeps seg faulting.

But prelim tests with wimax-range.cc and wimax-video-streamer.cc give a range of 750m and 8000m on the uplink and downlink respectively. All in all around 750 m as a simulation area should be workable.